% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_api.R
\name{autocomplete}
\alias{autocomplete}
\title{Base function for interacting with \code{autocomplete} of sections.}
\usage{
autocomplete(section, input, ...)
}
\arguments{
\item{section}{Defines what section of the api should be called.
\code{available_sections()} exposes all options for this parameter.}

\item{input}{The input to autocomplete. Could be a name, a place and similar.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> input query parameters.}
}
\value{
Returns the autocomplete information for the provided \code{section}
given the input.
}
\description{
This function is very handy when needing to either validate or
simply "fill in gaps" with autocompletion from the API.
}
\examples{
if (connection_check()) {
  autocomplete("regioner", "midt")
}
}
