% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_sections.R
\name{available_sections}
\alias{available_sections}
\title{List of available sections for \code{dawa()}}
\usage{
available_sections(as_list = FALSE)
}
\arguments{
\item{as_list}{This options controls the output format. The output is either
a named list or an atomic vector with the available sections}
}
\value{
The function returns either a named list or an atomic vector with the
available sections in the API. The format is decided by \code{as_list}.
}
\description{
Simple function to list all available options for the \code{section}
parameter in \code{dawa()}. Useful when getting to know the base api function of
this package.
}
\examples{

available_sections()

available_sections(as_list = TRUE)

}
