% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_check.R
\name{connection_check}
\alias{connection_check}
\title{Check that \strong{R} has access to resources at \verb{https://api.dataforsyningen.dk}}
\usage{
connection_check()
}
\value{
Returns a logical
}
\description{
Check that \strong{R} has access to resources at \verb{https://api.dataforsyningen.dk}
}
\examples{
\donttest{
connection_check()
}

}
\seealso{
Other Connection checks:
\code{\link[=status_check]{status_check()}}
}
\concept{Connection checks}
