% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_api.R
\name{dawa}
\alias{dawa}
\title{Base API function}
\usage{
dawa(
  section,
  ...,
  append_to_url = "",
  verbose = TRUE,
  cache = TRUE,
  dry_run = FALSE
)
}
\arguments{
\item{section}{Defines what section of the api should be called.
\code{available_sections()} exposes all options for this parameter.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> input query parameters.}

\item{append_to_url}{This options enables you to append a string to the url.
It is inserted in the request after the \code{section} but before the \code{...}.}

\item{verbose}{Wether or not to provide {cli} output for the user. Enabled by
default, but disabled for all functions that wrap \code{dawa()} in this package.}

\item{cache}{Option to use caching in a \code{tempdir()}. Enabled by default. More
info can be found in the documentation for \code{httr2::req_cache()}.}

\item{dry_run}{With this option enabled, the function will output the request
that has been created - Without sending it to the api. This is useful for
debugging.}
}
\description{
This function is the foundation of all communication with the
api in this package. Not intended for end-users, but it provides all the
flexibility of the api without the nice conveniences of the other functions
in this package.
}
\examples{
x <- dawa(section = "sogne")

x[[1]]
}
