% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from section}
\usage{
get_data(section, as_list = FALSE, ...)
}
\arguments{
\item{section}{Defines what section of the api should be called.
\code{available_sections()} exposes all options for this parameter.}

\item{as_list}{Whether or not to return the data as a list or in a dataframe.
Data.frame is the default output.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> input query parameters.}
}
\value{
Returns either a dataframe with all the data for the specified
section or a list with the same information.
}
\description{
Get data from section
}
\examples{
if (connection_check()) {
  x <- get_data("regioner")
  head(x)
}
}
\seealso{
Other Data functions: 
\code{\link{get_map_data}()}
}
\concept{Data functions}
