% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_data.R
\name{get_map_data}
\alias{get_map_data}
\title{Fetch map data to use with \code{{sf}}.}
\usage{
get_map_data(type)
}
\arguments{
\item{type}{Defines the type of map data to request from DAWA. Run
\code{available_sections(format = "geojson")} to your options.}
}
\value{
Returns a \code{data.frame} object that contains polygons for the section
provided.
}
\description{
Fetch map data to use with \code{{sf}}.
}
\section{The returned dataframe has date-time columns to inform when the information}{
was last changed. It is returned in Europe/Copenhagen time as that is the
\emph{local time} for the API. You can change that behavior by setting
\code{Sys.setenv(TZ = "Your/Timezone")}.
}

\examples{
x <- get_map_data("regioner")
ggplot2::ggplot(x) +
  ggplot2::geom_sf()
}
